project/requirements.txt
mysql-connector
Flask
Flask-MySQL